Various code smells
