+++ packages/web-service/src/pages/contact/applicant/applicant-address-form.js
@@ -1,7 +1,8 @@
-import { AccountRoles, checkHasContact, ContactRoles } from '../common/common.js'
+import { checkHasContact } from '../common/common.js'
 import { addressFormPage } from '../common/address-form/address-form-page.js'
 import { contactURIs } from '../../../uris.js'
 import { getAddressFormData, setAddressFormData } from '../common/address-form/address-form.js'
+import { AccountRoles, ContactRoles } from '../common/contact-roles.js'
 
 const { ADDRESS_FORM, CHECK_ANSWERS } = contactURIs.APPLICANT
