+++ packages/web-service/src/pages/contact/applicant/applicant-address.js
@@ -1,7 +1,8 @@
-import { AccountRoles, checkAccountAndOrContactData, ContactRoles } from '../common/common.js'
+import { checkAccountAndOrContactData } from '../common/common.js'
 import { addressPage } from '../common/address/address-page.js'
 import { contactURIs } from '../../../uris.js'
 import { getAddressData, setAddressData } from '../common/address/address.js'
+import { AccountRoles, ContactRoles } from '../common/contact-roles.js'
 
 const { ADDRESS, CHECK_ANSWERS } = contactURIs.APPLICANT
