+++ packages/web-service/src/pages/contact/applicant/applicant-check-answers.js
@@ -1,7 +1,8 @@
 import { contactURIs, TASKLIST } from '../../../uris.js'
 import { getCheckAnswersData } from '../common/check-answers/check-answers.js'
 import { checkAnswersPage } from '../../common/check-answers.js'
-import { AccountRoles, checkHasApplication, ContactRoles } from '../common/common.js'
+import { checkHasApplication } from '../common/common.js'
+import { AccountRoles, ContactRoles } from '../common/contact-roles.js'
 const { CHECK_ANSWERS } = contactURIs.APPLICANT
 
 export const applicantCheckAnswers = checkAnswersPage({
