+++ packages/web-service/src/pages/contact/applicant/applicant-name.js
@@ -1,7 +1,8 @@
 import { contactURIs } from '../../../uris.js'
 import { contactNamePage } from '../common/contact-name/contact-name-page.js'
 import { contactNameCompletion, getContactData, setContactData } from '../common/contact-name/contact-name.js'
-import { AccountRoles, checkHasContact, ContactRoles } from '../common/common.js'
+import { checkHasContact } from '../common/common.js'
+import { AccountRoles, ContactRoles } from '../common/contact-roles.js'
 
 const { NAME } = contactURIs.APPLICANT
