+++ packages/web-service/src/pages/contact/applicant/applicant-organisation.js
@@ -4,8 +4,9 @@ import {
   contactAccountCompletion
 } from '../common/is-organisation/is-organisation.js'
 import { contactURIs } from '../../../uris.js'
-import { AccountRoles, checkHasContact, ContactRoles } from '../common/common.js'
+import { checkHasContact } from '../common/common.js'
 import { isOrganisation } from '../common/is-organisation/is-organisation-page.js'
+import { AccountRoles, ContactRoles } from '../common/contact-roles.js'
 
 const { IS_ORGANISATION } = contactURIs.APPLICANT
