+++ packages/web-service/src/pages/contact/applicant/applicant-organisations.js
@@ -1,7 +1,7 @@
 import { accountNamesPage } from '../common/account-names/account-names-page.js'
 import { contactURIs } from '../../../uris.js'
 import { getAccountNamesData, setAccountNamesData, accountNamesCompletion, accountNamesCheckData } from '../common/account-names/account-names.js'
-import { AccountRoles, ContactRoles } from '../common/common.js'
+import { AccountRoles, ContactRoles } from '../common/contact-roles.js'
 const { ORGANISATIONS } = contactURIs.APPLICANT
 
 export const applicantOrganisations = accountNamesPage({
