+++ packages/web-service/src/pages/contact/applicant/applicant-postcode.js
@@ -1,7 +1,8 @@
 import { contactURIs } from '../../../uris.js'
 import { postcodePage } from '../common/postcode/postcode-page.js'
 import { getPostcodeData, postcodeCompletion, setPostcodeData } from '../common/postcode/postcode.js'
-import { AccountRoles, checkHasContact, ContactRoles } from '../common/common.js'
+import { checkHasContact } from '../common/common.js'
+import { AccountRoles, ContactRoles } from '../common/contact-roles.js'
 
 const { POSTCODE } = contactURIs.APPLICANT
 
@@ -10,6 +11,6 @@ export const applicantPostcode = postcodePage({
   uri: POSTCODE.uri,
   checkData: checkHasContact(ContactRoles.APPLICANT, contactURIs.APPLICANT),
   getData: getPostcodeData(ContactRoles.APPLICANT, AccountRoles.APPLICANT_ORGANISATION, contactURIs.APPLICANT),
-  setData: setPostcodeData(ContactRoles.APPLICANT, AccountRoles.APPLICANT_ORGANISATION),
+  setData: setPostcodeData(ContactRoles.APPLICANT),
   completion: postcodeCompletion(contactURIs.APPLICANT)
 })
