+++ packages/web-service/src/pages/contact/applicant/applicant-user.js
@@ -1,8 +1,9 @@
 import { contactURIs } from '../../../uris.js'
 import { getUserData, setUserData, userCompletion } from '../common/user/user.js'
-import { AccountRoles, checkHasApplication, ContactRoles } from '../common/common.js'
+import { checkHasApplication } from '../common/common.js'
 
 import { yesNoPage } from '../../common/yes-no.js'
+import { AccountRoles, ContactRoles } from '../common/contact-roles.js'
 const { USER } = contactURIs.APPLICANT
 
 export const applicantUser = yesNoPage({
