+++ packages/web-service/src/pages/contact/authorised-people/authorised-person-email.js
@@ -1,7 +1,8 @@
 import { contactURIs } from '../../../uris.js'
 import { emailAddressPage } from '../common/email-address/email-address-page.js'
-import { contactAccountOperationsForContactAccount, ContactRoles } from '../common/common.js'
+import { contactAccountOperationsForContactAccount } from '../common/common.js'
 import { checkAuthorisedPeopleData, getAuthorisedPeopleCompletion, getAuthorisedPeopleData } from './common.js'
+import { ContactRoles } from '../common/contact-roles.js'
 
 const { EMAIL } = contactURIs.AUTHORISED_PEOPLE
