+++ packages/web-service/src/pages/contact/authorised-people/remove-authorised-person.js
@@ -1,7 +1,8 @@
 import { yesNoPage } from '../../common/yes-no.js'
 import { contactURIs } from '../../../uris.js'
 import { checkAuthorisedPeopleData, getAuthorisedPeopleData } from './common.js'
-import { contactOperationsForContact, ContactRoles } from '../common/common.js'
+import { contactOperationsForContact } from '../common/common.js'
+import { ContactRoles } from '../common/contact-roles.js'
 
 const { ADD, REMOVE } = contactURIs.AUTHORISED_PEOPLE
