+++ packages/web-service/src/server.js
@@ -163,7 +163,7 @@ const init = async server => {
   }
 
   // Log any errors
-  // server.ext('onPreResponse', errorHandler)
+  server.ext('onPreResponse', errorHandler)
 
   /*
    * Set up shutdown handlers
